refactor: move code to reactiveEffect.ts
